extract mocks to their own directory
