+++ WooCommerce/Classes/POS/TabBar/POSTabCoordinator.swift
@@ -7,10 +7,8 @@ import class WooFoundation.CurrencySettings
 import WooFoundationCore
 import protocol Storage.GRDBManagerProtocol
 import protocol Storage.StorageManagerType
-import class WooFoundationCore.CurrencyFormatter
 import struct NetworkingCore.JetpackSite
 import struct NetworkingCore.OrderItem
-import struct Combine.AnyPublisher
 import PointOfSale
 
 protocol POSTabVisibilityCheckerProtocol {
@@ -314,261 +312,5 @@ private extension POSTabCoordinator {
     /// Decorates track events with a different prefix when Point of Sale is active.
     func updateTrackEventPrefix(_ isPointOfSaleActive: Bool) {
         TracksProvider.setPOSMode(isPointOfSaleActive)
-    }
-}
-
-/// Mock order service for screenshot tests that returns immediate loaded state
-private final class POSOrderServiceScreenshotMock: POSOrderServiceProtocol {
-    private let currency: String
-
-    init(currency: String) {
-        self.currency = currency
-    }
-
-    func syncOrder(cart: POSCart, currency: CurrencyCode) async throws -> Order {
-        // Create a mock order with totals calculated from the cart
-        // For screenshot tests with 2 products: $35.00 + $45.00 = $80.00
-        let orderItems = [
-            OrderItem(
-                itemID: 1,
-                name: "Product 1",
-                productID: 1,
-                variationID: 0,
-                quantity: 1,
-                price: NSDecimalNumber(string: "35.00"),
-                sku: nil,
-                subtotal: "35.00",
-                subtotalTax: "0.00",
-                taxClass: "",
-                taxes: [],
-                total: "35.00",
-                totalTax: "0.00",
-                attributes: [],
-                addOns: [],
-                image: nil,
-                parent: nil,
-                bundleConfiguration: []
-            ),
-            OrderItem(
-                itemID: 2,
-                name: "Product 2",
-                productID: 2,
-                variationID: 0,
-                quantity: 1,
-                price: NSDecimalNumber(string: "45.00"),
-                sku: nil,
-                subtotal: "45.00",
-                subtotalTax: "0.00",
-                taxClass: "",
-                taxes: [],
-                total: "45.00",
-                totalTax: "0.00",
-                attributes: [],
-                addOns: [],
-                image: nil,
-                parent: nil,
-                bundleConfiguration: []
-            )
-        ]
-
-        let total = "80.00"
-        let tax = "0.00"
-
-        return Order(siteID: 0,
-                    orderID: 1,
-                    parentID: 0,
-                    customerID: 0,
-                    orderKey: "",
-                    isEditable: true,
-                    needsPayment: true,
-                    needsProcessing: true,
-                    number: "1",
-                    status: .pending,
-                    currency: currency.rawValue,
-                    currencySymbol: "$",
-                    customerNote: nil,
-                    dateCreated: Date(),
-                    dateModified: Date(),
-                    datePaid: nil,
-                    discountTotal: "0.00",
-                    discountTax: "0.00",
-                    shippingTotal: "0.00",
-                    shippingTax: "0.00",
-                    total: total,
-                    totalTax: tax,
-                    paymentMethodID: "",
-                    paymentMethodTitle: "",
-                    paymentURL: nil,
-                    chargeID: nil,
-                    items: orderItems,  // Necessary for the card payment flow to be presented in POS
-                    billingAddress: nil,
-                    shippingAddress: nil,
-                    shippingLines: [],
-                    coupons: [],
-                    refunds: [],
-                    fees: [],
-                    taxes: [],
-                    customFields: [],
-                    renewalSubscriptionID: nil,
-                    appliedGiftCards: [],
-                    attributionInfo: nil,
-                    shippingLabels: [],
-                    createdVia: "pos")
-    }
-
-    func updatePOSOrder(orderID: Int64, recipientEmail: String) async throws {}
-
-    func markOrderAsCompletedWithCashPayment(order: Order, changeDueAmount: String?) async throws {}
-}
-
-
-private final class PointOfSaleItemServiceScreenshotMock: Yosemite.PointOfSaleItemServiceProtocol {
-
-    func providePointOfSaleItems(pageNumber: Int,
-                                 fetchStrategy: Yosemite.PointOfSalePurchasableItemFetchStrategy) async throws -> PagedItems<Yosemite.POSItem> {
-        let port = UserDefaults.standard.integer(forKey: "mocks-port")
-        let mockResourceUrlHost = "http://localhost:\(port)/"
-
-        let mockItems = Self.makeScreenshotMockItems(mockResourceUrlHost: mockResourceUrlHost)
-
-        return PagedItems(items: mockItems, hasMorePages: false, totalItems: mockItems.count)
-    }
-
-    func providePointOfSaleVariationItems(for parentProduct: Yosemite.POSVariableParentProduct,
-                                          pageNumber: Int,
-                                          fetchStrategy: Yosemite.PointOfSalePurchasableItemFetchStrategy) async throws -> PagedItems<Yosemite.POSItem> {
-        // Not needed for screenshot tests, return empty
-        return PagedItems(items: [], hasMorePages: false, totalItems: 0)
-    }
-
-    private static func makeScreenshotMockItems(mockResourceUrlHost: String) -> [Yosemite.POSItem] {
-        let product1 = Yosemite.POSSimpleProduct(
-            id: UUID(),
-            name: "Rose Gold Shades",
-            formattedPrice: "$35.00",
-            productImageSource: mockResourceUrlHost + "rose-gold-shades",
-            productID: 1,
-            price: "35.00",
-            manageStock: false,
-            stockQuantity: nil,
-            stockStatusKey: "instock"
-        )
-
-        let product2 = Yosemite.POSSimpleProduct(
-            id: UUID(),
-            name: "Black Coral Shades",
-            formattedPrice: "$45.00",
-            productImageSource: mockResourceUrlHost + "black-coral-shades",
-            productID: 2,
-            price: "45.00",
-            manageStock: false,
-            stockQuantity: nil,
-            stockStatusKey: "instock"
-        )
-
-        let product3 = Yosemite.POSSimpleProduct(
-            id: UUID(),
-            name: "Akoya Pearl Shades",
-            formattedPrice: "$50.00",
-            productImageSource: mockResourceUrlHost + "akoya-pearl-shades",
-            productID: 3,
-            price: "50.00",
-            manageStock: true,
-            stockQuantity: 10,
-            stockStatusKey: "instock"
-        )
-
-        let product4 = Yosemite.POSSimpleProduct(
-            id: UUID(),
-            name: "Malaya Shades",
-            formattedPrice: "$40.00",
-            productImageSource: mockResourceUrlHost + "malaya-shades",
-            productID: 4,
-            price: "40.00",
-            manageStock: false,
-            stockQuantity: nil,
-            stockStatusKey: "instock"
-        )
-
-        return [
-            .simpleProduct(product1),
-            .simpleProduct(product2),
-            .simpleProduct(product3),
-            .simpleProduct(product4)
-        ]
-    }
-}
-
-final class CardPresentPaymentServiceScreenshotMock: CardPresentPaymentFacade {
-    let paymentEventPublisher: AnyPublisher<CardPresentPaymentEvent, Never>
-    let readerConnectionStatusPublisher: AnyPublisher<CardPresentPaymentReaderConnectionStatus, Never>
-    let cardReaderUpdateStatePublisher: AnyPublisher<CardReaderSoftwareUpdateState, Never>
-
-    private let paymentEventSubject = PassthroughSubject<CardPresentPaymentEvent, Never>()
-
-    init() {
-        paymentEventPublisher = paymentEventSubject
-            .receive(on: DispatchQueue.main)
-            .eraseToAnyPublisher()
-
-        // Always return a connected card reader for screenshots
-        let mockReader = CardPresentPaymentCardReader(
-            name: "Simulated POS E",
-            batteryLevel: 0.5,
-            softwareVersion: "1.00.03.34-SZZZ_Generic_v45-300001"
-        )
-        readerConnectionStatusPublisher = Just(.connected(mockReader))
-            .receive(on: DispatchQueue.main)
-            .eraseToAnyPublisher()
-
-        // No updates needed for screenshots
-        cardReaderUpdateStatePublisher = Just(.none)
-            .receive(on: DispatchQueue.main)
-            .eraseToAnyPublisher()
-    }
-
-    func connectReader(using connectionMethod: CardReaderConnectionMethod) async throws -> CardPresentPaymentReaderConnectionResult {
-        // Return connected reader immediately
-        let mockReader = CardPresentPaymentCardReader(
-            name: "Simulated POS E",
-            batteryLevel: 0.5,
-            softwareVersion: "1.00.03.34-SZZZ_Generic_v45-300001"
-        )
-        return .connected(mockReader)
-    }
-
-    func disconnectReader() async {
-        // No-op for screenshots
-    }
-
-    func updateCardReaderSoftware() async throws {
-        // No-op for screenshots
-    }
-
-    func collectPayment(for order: Order, using connectionMethod: CardReaderConnectionMethod, channel: PaymentChannel) async throws -> CardPresentPaymentResult {
-
-        // 1. Validating order
-        paymentEventSubject.send(.show(eventDetails: .validatingOrder(cancelPayment: {})))
-        try await Task.sleep(nanoseconds: 100_000_000) // 0.1 seconds
-
-        // 2. Preparing reader
-        paymentEventSubject.send(.show(eventDetails: .preparingForPayment(cancelPayment: {})))
-        try await Task.sleep(nanoseconds: 100_000_000) // 0.1 seconds
-
-        // 3. Ready to accept card
-        let inputMethods: Yosemite.CardReaderInput = [.tap, .swipe, .insert]
-        paymentEventSubject.send(.show(eventDetails: .tapSwipeOrInsertCard(inputMethods: inputMethods, cancelPayment: {})))
-
-        try await Task.sleep(nanoseconds: 3_000_000_000) // 3 seconds, give it some time for the screenshot
-
-        return .success(CardPresentPaymentTransaction())
-    }
-
-    func cancelPayment() {
-        // No-op for screenshots
-    }
-
-    func cancelPayment() async throws {
-        // No-op for screenshots
     }
 }
